+++ xcm_integration_tests/emulated/chains/relays/dancelight/Cargo.toml
@@ -14,7 +14,6 @@ workspace = true
 [dependencies]
 dancelight-runtime = { workspace = true }
 dancelight-runtime-constants = { workspace = true }
-dancelight-runtime-test-utils = { workspace = true }
 emulated-integration-tests-common = { workspace = true }
 runtime-parachains = { workspace = true }
 sp-core = { workspace = true }
