+++ xcm_integration_tests/emulated/common/Cargo.toml
@@ -32,6 +32,8 @@ xcm-emulator = { workspace = true, default-features = true }
 dancelight-runtime = { workspace = true }
 keyring = { workspace = true }
 snowbridge-pallet-outbound-queue = { workspace = true }
+snowbridge-beacon-primitives = { workspace = true }
+snowbridge-core = { workspace = true }
 
 [features]
 default = [ "std" ]
@@ -44,6 +46,8 @@ std = [
 	"keyring/std",
 	"pallet-message-queue/std",
 	"snowbridge-pallet-outbound-queue/std",
+	"snowbridge-beacon-primitives/std",
+	"snowbridge-core/std",
 	"sp-core/std",
 	"sp-std/std",
 	"xcm/std",
