--- xcm_integration_tests/emulated/dancelight-runtime-test-utils/Cargo.toml
@@ -1,113 +1,0 @@
-[package]
-name = "dancelight-runtime-test-utils"
-description = "Dancelight runtime test utils"
-edition = "2021"
-license = "GPL-3.0-only"
-publish = false
-version = "0.0.0"
-[package.authors]
-workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-bitvec = { workspace = true }
-parity-scale-codec = { workspace = true, features = [ "derive", "max-encoded-len" ] }
-scale-info = { workspace = true, features = [ "derive" ] }
-
-babe-primitives = { workspace = true }
-beefy-primitives = { workspace = true }
-dancelight-runtime = { workspace = true }
-frame-support = { workspace = true, features = [ "tuples-96" ] }
-frame-system = { workspace = true }
-grandpa_primitives = { workspace = true }
-nimbus-primitives = { workspace = true }
-pallet-babe = { workspace = true }
-pallet-balances = { workspace = true }
-pallet-external-validators = { workspace = true }
-pallet-grandpa = { workspace = true }
-pallet-invulnerables = { workspace = true }
-pallet-session = { workspace = true }
-pallet-sudo = { workspace = true }
-pallet-xcm = { workspace = true }
-primitives = { workspace = true }
-runtime-common = { workspace = true }
-runtime-parachains = { workspace = true }
-sp-core = { workspace = true }
-sp-io = { workspace = true }
-sp-keystore = { workspace = true }
-sp-runtime = { workspace = true }
-sp-std = { workspace = true }
-sp-storage = { workspace = true }
-
-# Cumulus
-cumulus-primitives-core = { workspace = true }
-
-# Tanssi
-tp-data-preservers-common = { workspace = true }
-tp-traits = { workspace = true }
-
-pallet-author-noting = { workspace = true }
-pallet-configuration = { workspace = true }
-pallet-data-preservers = { workspace = true }
-pallet-registrar = { workspace = true }
-pallet-registrar-runtime-api = { workspace = true }
-pallet-services-payment = { workspace = true }
-
-# Bridges
-finality-grandpa = { workspace = true, default-features = true, features = [ "derive-codec" ] }
-milagro-bls = { workspace = true, features = [ "std" ] }
-rand = { workspace = true, features = [ "std", "std_rng" ] }
-snowbridge-beacon-primitives = { workspace = true }
-snowbridge-core = { workspace = true }
-snowbridge-pallet-system = { workspace = true }
-test-relay-sproof-builder = { workspace = true }
-
-[features]
-default = [ "std" ]
-std = [
-	"babe-primitives/std",
-	"beefy-primitives/std",
-	"bitvec/std",
-	"cumulus-primitives-core/std",
-	"dancelight-runtime/std",
-	"frame-support/std",
-	"frame-system/std",
-	"grandpa_primitives/std",
-	"milagro-bls/std",
-	"nimbus-primitives/std",
-	"pallet-author-noting/std",
-	"pallet-babe/std",
-	"pallet-balances/std",
-	"pallet-configuration/std",
-	"pallet-data-preservers/std",
-	"pallet-external-validators/std",
-	"pallet-grandpa/std",
-	"pallet-invulnerables/std",
-	"pallet-registrar-runtime-api/std",
-	"pallet-registrar/std",
-	"pallet-services-payment/std",
-	"pallet-session/std",
-	"pallet-sudo/std",
-	"pallet-xcm/std",
-	"parity-scale-codec/std",
-	"primitives/std",
-	"rand/std",
-	"runtime-common/std",
-	"runtime-parachains/std",
-	"scale-info/std",
-	"snowbridge-beacon-primitives/std",
-	"snowbridge-core/std",
-	"snowbridge-pallet-system/std",
-	"sp-core/std",
-	"sp-io/std",
-	"sp-keystore/std",
-	"sp-runtime/std",
-	"sp-std/std",
-	"sp-storage/std",
-	"test-relay-sproof-builder/std",
-	"tp-data-preservers-common/std",
-	"tp-traits/std",
-]
-no_std = []
