+++ xcm_integration_tests/emulated/tests/assets/dancelight/Cargo.toml
@@ -12,10 +12,10 @@ workspace = true
 workspace = true
 
 [dependencies]
+tanssi-emulated-integration-tests-common = { workspace = true }
 container-chain-template-simple-runtime = { workspace = true }
 dancelight-emulated-chain = { workspace = true }
 dancelight-runtime = { workspace = true }
-dancelight-runtime-test-utils = { workspace = true }
 dancelight-system-emulated-network  = { workspace = true }
 frame-support = { workspace = true, features = [ "tuples-96" ] }
 pallet-message-queue = { workspace = true }
