+++ xcm_integration_tests/emulated/tests/bridges/dancelight/src/tests/mod.rs
@@ -23,7 +23,6 @@ mod imports {
         alloy_sol_types::SolEvent,
         dancelight_emulated_chain::DancelightRelayPallet,
         dancelight_runtime::Runtime,
-        dancelight_runtime_test_utils::mock_snowbridge_message_proof,
         dancelight_system_emulated_network::DancelightRelay as Dancelight,
         dancelight_system_emulated_network::DancelightSender,
         dancelight_system_emulated_network::SimpleTemplateDancelightPara as DancelightPara,
@@ -36,6 +35,7 @@ mod imports {
         },
         snowbridge_router_primitives::inbound::envelope::OutboundMessageAccepted,
         sp_core::H256,
+        tanssi_emulated_integration_tests_common::snowbridge::mock_snowbridge_message_proof,
         tp_bridge::{
             symbiotic_message_processor::{
                 InboundCommand, Message as SymbioticMessage, Payload, MAGIC_BYTES,
